Terrain3D/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "terrain3d"
version = "0.1.0"
description = "AI-powered terrain generation via diffusion models (terrain-diffusion)"
readme = "README.md"
authors = [{ name = "Terrain3D Project" }]
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "numpy",
    "pillow",
    "tifffile",
    "click>=8.1.0",
    "rich>=13.0.0",
    "rich-click>=1.8.0",
    "gamedev-shared @ file:../Shared",
    # AI / diffusion deps
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "diffusers>=0.30.2",
    "transformers>=4.44.0",
    "accelerate>=0.34.0",
    "safetensors>=0.4.5",
    "infinite-tensor>=0.3.0",
    "h5py>=3.12.0",
    "scipy>=1.10.0",
    "scikit-image>=0.24.0",
    "pyfastnoiselite==0.0.6",
    "rasterio>=1.4.0",
    "numba>=0.59.0",
    "matplotlib>=3.9.0",
    "ema-pytorch>=0.7.0",
    "easydict>=1.10",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Topic :: Multimedia :: Graphics :: 3D Modeling",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-cov>=4.0.0", "ruff>=0.11.0"]

[project.scripts]
terrain3d = "terrain3d.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
